# Optimize MultiPart and ParallelHatch performance