--- a/src/main/java/com/gregtechceu/gtceu/client/renderer/machine/HPCAPartRenderer.java
+++ b/src/main/java/com/gregtechceu/gtceu/client/renderer/machine/HPCAPartRenderer.java
@@ -70,7 +70,7 @@ public class HPCAPartRenderer extends TieredHullMachineRenderer {
         super.renderMachine(quads, definition, machine, frontFacing, side, rand, modelFacing, modelState);
         if (machine instanceof HPCAComponentPartMachine hpcaComponent) {
             ResourceLocation texture, emissiveTexture = null;
-            var controller = hpcaComponent.getControllers().isEmpty() ? null : hpcaComponent.getControllers().get(0);
+            var controller = hpcaComponent.isFormed() ? hpcaComponent.getControllers().first() : null;
             if (controller != null && (controller instanceof IWorkable workable && workable.isActive())) {
                 if (hpcaComponent.isDamaged()) {
                     texture = damagedActiveTexture;