--- a/src/main/java/com/gregtechceu/gtceu/common/machine/multiblock/part/OpticalDataHatchMachine.java
+++ b/src/main/java/com/gregtechceu/gtceu/common/machine/multiblock/part/OpticalDataHatchMachine.java
@@ -42,37 +42,39 @@ public class OpticalDataHatchMachine extends MultiblockPartMachine implements IO
     @Override
     public boolean isRecipeAvailable(@NotNull GTRecipe recipe, @NotNull Collection<IDataAccessHatch> seen) {
         seen.add(this);
-        if (!getControllers().isEmpty()) {
-            if (isTransmitter()) {
-                IMultiController controller = getControllers().get(0);
-                if (!(controller instanceof IWorkable workable) || !workable.isActive()) return false;
+        if (!isFormed()) {
+            return false;
+        }
 
-                List<IDataAccessHatch> dataAccesses = new ArrayList<>();
-                List<IDataAccessHatch> transmitters = new ArrayList<>();
-                for (var part : controller.getParts()) {
-                    Block block = part.self().getBlockState().getBlock();
-                    if (part instanceof IDataAccessHatch hatch && PartAbility.DATA_ACCESS.isApplicable(block)) {
-                        dataAccesses.add(hatch);
-                    }
-                    if (part instanceof IDataAccessHatch hatch &&
-                            PartAbility.OPTICAL_DATA_RECEPTION.isApplicable(block)) {
-                        transmitters.add(hatch);
-                    }
+        if (isTransmitter()) {
+            IMultiController controller = getControllers().first();
+            if (!(controller instanceof IWorkable workable) || !workable.isActive()) return false;
+
+            List<IDataAccessHatch> dataAccesses = new ArrayList<>();
+            List<IDataAccessHatch> transmitters = new ArrayList<>();
+            for (var part : controller.getParts()) {
+                Block block = part.self().getBlockState().getBlock();
+                if (part instanceof IDataAccessHatch hatch && PartAbility.DATA_ACCESS.isApplicable(block)) {
+                    dataAccesses.add(hatch);
                 }
-
-                return isRecipeAvailable(dataAccesses, seen, recipe) ||
-                        isRecipeAvailable(transmitters, seen, recipe);
-            } else {
-                BlockEntity tileEntity = getLevel().getBlockEntity(getPos().relative(getFrontFacing()));
-                if (tileEntity == null) return false;
-
-                if (tileEntity instanceof OpticalPipeBlockEntity) {
-                    // noinspection DataFlowIssue
-                    IDataAccessHatch cap = tileEntity.getCapability(GTCapability.CAPABILITY_DATA_ACCESS,
-                            getFrontFacing().getOpposite()).orElse(null);
-                    // noinspection ConstantValue
-                    return cap != null && cap.isRecipeAvailable(recipe, seen);
+                if (part instanceof IDataAccessHatch hatch &&
+                        PartAbility.OPTICAL_DATA_RECEPTION.isApplicable(block)) {
+                    transmitters.add(hatch);
                 }
+            }
+
+            return isRecipeAvailable(dataAccesses, seen, recipe) ||
+                    isRecipeAvailable(transmitters, seen, recipe);
+        } else {
+            BlockEntity tileEntity = getLevel().getBlockEntity(getPos().relative(getFrontFacing()));
+            if (tileEntity == null) return false;
+
+            if (tileEntity instanceof OpticalPipeBlockEntity) {
+                // noinspection DataFlowIssue
+                IDataAccessHatch cap = tileEntity.getCapability(GTCapability.CAPABILITY_DATA_ACCESS,
+                        getFrontFacing().getOpposite()).orElse(null);
+                // noinspection ConstantValue
+                return cap != null && cap.isRecipeAvailable(recipe, seen);
             }
         }
         return false;