--- a/src/main/java/com/gregtechceu/gtceu/common/machine/multiblock/part/RotorHolderPartMachine.java
+++ b/src/main/java/com/gregtechceu/gtceu/common/machine/multiblock/part/RotorHolderPartMachine.java
@@ -115,6 +115,11 @@ public class RotorHolderPartMachine extends TieredPartMachine
         }
     }
 
+    @Override
+    public boolean canShared() {
+        return false;
+    }
+
     //////////////////////////////////////
     // ****** Rotor Holder ******//
     //////////////////////////////////////
@@ -144,12 +149,8 @@ public class RotorHolderPartMachine extends TieredPartMachine
     }
 
     private void updateRotorSpeed() {
-        for (IMultiController controller : getControllers()) {
-            if (controller instanceof IWorkableMultiController workableMultiController) {
-                if (workableMultiController.getRecipeLogic().isWorking()) {
-                    return;
-                }
-            }
+        if (isFormed() && getControllers().first() instanceof IWorkableMultiController workable) {
+            if (workable.getRecipeLogic().isWorking()) return;
         }
         if (!hasRotor()) {
             setRotorSpeed(0);
@@ -167,11 +168,8 @@ public class RotorHolderPartMachine extends TieredPartMachine
         }
         if (self().getOffsetTimer() % 20 == 0) {
             var numMaintenanceProblems = 0;
-            for (IMultiPart part : controller.getParts()) {
-                if (part instanceof IMaintenanceMachine maintenance) {
-                    numMaintenanceProblems = maintenance.getNumMaintenanceProblems();
-                    break;
-                }
+            if (isFormed() && getControllers().first() instanceof IMaintenanceMachine maintenance) {
+                numMaintenanceProblems = maintenance.getNumMaintenanceProblems();
             }
             damageRotor(1 + numMaintenanceProblems);
         }
@@ -179,10 +177,8 @@ public class RotorHolderPartMachine extends TieredPartMachine
     }
 
     public int getTierDifference() {
-        for (IMultiController controller : getControllers()) {
-            if (controller instanceof ITieredMachine tieredMachine) {
-                return getTier() - tieredMachine.getTier();
-            }
+        if (isFormed() && getControllers().first() instanceof ITieredMachine tieredMachine) {
+            return getTier() - tieredMachine.getTier();
         }
         return -1;
     }