## What
This PR aims to optimize some parts of the Multiblock Lifecycle. 
1.  Whenever an `IMultiPart` needed to query the controllers it was attached to, it would iterate through all of its known controller positions and return a new List every time. 
2. When the current parallel setting of a multiblock is needed, its parts list would be searched for the parallel hatch *every time*, which is especially noticeable when using many GCYM machines.

## Implementation Details
1. As it is highly unlikely for an `IMultiController` to be moved from its current position without triggering a structure update and thus removing itself from the controller set, we can also keep track of a Set of `IMultiControllers`. I have used a `ReferenceLinkedOpenHashSet` from FastUtil, as we only really care about references and using a LinkedSet provides us with consistent iterations, as well as quick access to the first controller in the set.
This Set does need to be populated manually on the client-side due to an inability to sync, so that is the reason why maintaining the `controllerPositions` set is useful.
2. It is very unlikely that the parallel hatch of a multiblock will change while it is formed, and thus can be cached. This has been implemented in the `MultiblockControllerMachine` class as a simple field. I have chosen to make the method return an `Optional` to better fit in with our current usages, but it can be switched to a `Nullable` if desired.

Additionally, some small tweaks have been added:
- RotorHolderPartMachines now cannot be shared and its methods have been updated to reflect that
- `!getControllers.isEmpty()` calls have been changed to `isFormed()`
## Outcome

Performance. 

600s runs in fairly stressful situations
`IMultiPart`
| Before| After|
|--------|--------|
| ![image](https://github.com/user-attachments/assets/8375f9b1-c157-415a-abdd-3b21ed1bcb6f) | ![image](https://github.com/user-attachments/assets/cf78a3e9-6112-45a4-b91f-c12a4c14f612) |

`IParallelHatch`
TBD